requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
DateTime==5.1
Django==4.2.2
djangorestframework==3.14.0
idna==3.4
mysql-connector-python==8.0.33
protobuf==3.20.3
pytz==2023.3
requests==2.31.0
schedule==1.2.0
sqlparse==0.4.4
urllib3==2.0.3
zope.interface==6.0
